requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
groq==0.4.1
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1

